Updated Files

+++ ParentsBank/Controllers/AccountDetailsController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
 using System.Linq;
+using System.Threading.Tasks;
 using System.Net;
 using System.Web;
 using System.Web.Mvc;
@@ -15,19 +16,20 @@ namespace ParentsBank.Controllers
         private ApplicationDbContext db = new ApplicationDbContext();
 
         // GET: AccountDetails
-        public ActionResult Index()
+        public async Task<ActionResult> Index()
         {
-            return View(db.Accounts.ToList());
+            string user = User.Identity.Name;
+            return View(await db.Accounts.Where(model => model.Owner.ToLower() == user.ToLower()).Union(db.Accounts.Where(model => model.Recipient.ToLower() == user.ToLower())).ToListAsync());
         }
 
         // GET: AccountDetails/Details/5
-        public ActionResult Details(int? id)
+        public async Task<ActionResult> Details(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            AccountDetails accountDetails = db.Accounts.Find(id);
+            AccountDetails accountDetails = await db.Accounts.FindAsync(id);
             if (accountDetails == null)
             {
                 return HttpNotFound();
@@ -46,12 +48,15 @@ namespace ParentsBank.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "Id,Owner,Recipient,Name,OpenDate,InterestRate")] AccountDetails accountDetails)
+        public async Task<ActionResult> Create([Bind(Include = "Id,Owner,Recipient,Name,OpenDate,InterestRate,Balance")] AccountDetails accountDetails)
         {
+            accountDetails.OpenDate = DateTime.Now;
+            accountDetails.Owner = User.Identity.Name;
+            ValidateAccountDetails(accountDetails,0);
             if (ModelState.IsValid)
             {
                 db.Accounts.Add(accountDetails);
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
 
@@ -59,13 +64,13 @@ namespace ParentsBank.Controllers
         }
 
         // GET: AccountDetails/Edit/5
-        public ActionResult Edit(int? id)
+        public async Task<ActionResult> Edit(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            AccountDetails accountDetails = db.Accounts.Find(id);
+            AccountDetails accountDetails = await db.Accounts.FindAsync(id);
             if (accountDetails == null)
             {
                 return HttpNotFound();
@@ -78,25 +83,26 @@ namespace ParentsBank.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id,Owner,Recipient,Name,OpenDate,InterestRate")] AccountDetails accountDetails)
+        public async Task<ActionResult> Edit([Bind(Include = "Id,Owner,Recipient,Name,OpenDate,InterestRate,Balance")] AccountDetails accountDetails)
         {
+            ValidateAccountDetails(accountDetails,1);
             if (ModelState.IsValid)
             {
                 db.Entry(accountDetails).State = EntityState.Modified;
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
             return View(accountDetails);
         }
 
         // GET: AccountDetails/Delete/5
-        public ActionResult Delete(int? id)
+        public async Task<ActionResult> Delete(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            AccountDetails accountDetails = db.Accounts.Find(id);
+            AccountDetails accountDetails = await db.Accounts.FindAsync(id);
             if (accountDetails == null)
             {
                 return HttpNotFound();
@@ -107,12 +113,20 @@ namespace ParentsBank.Controllers
         // POST: AccountDetails/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public async Task<ActionResult> DeleteConfirmed(int id)
         {
-            AccountDetails accountDetails = db.Accounts.Find(id);
-            db.Accounts.Remove(accountDetails);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            AccountDetails accountDetails = await db.Accounts.FindAsync(id);
+            if (accountDetails.Balance == 0)
+            {
+                db.Accounts.Remove(accountDetails);
+                await db.SaveChangesAsync();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                ModelState.AddModelError("", "Cannot delete account with balance.");
+                return View(accountDetails);
+            }
         }
 
         protected override void Dispose(bool disposing)
@@ -122,6 +136,56 @@ namespace ParentsBank.Controllers
                 db.Dispose();
             }
             base.Dispose(disposing);
+        }
+
+        public bool CheckAccountRecipientEmail(string recipientEmail, int exCnt)
+        {
+            int count = db.Accounts.Where(acct => acct.Recipient.ToLower() == recipientEmail.ToLower()).Count();
+            if (count > exCnt)
+            {
+                return true;
+            }
+            return false;
+        }
+
+        public bool CheckAccountRecipientOwner(string recipientEmail)
+        {
+            int count = db.Accounts.Where(acct => acct.Owner.ToLower() == recipientEmail.ToLower()).Count();
+            if (count > 0)
+            {
+                return true;
+            }
+            return false;
+        }
+
+        public bool CheckAccountOwnerRecipient(string ownerEmail)
+        {
+            int count = db.Accounts.Where(acct => acct.Recipient.ToLower() == ownerEmail.ToLower()).Count();
+            if (count > 0)
+            {
+                return true;
+            }
+            return false;
+        }
+
+        public void ValidateAccountDetails(AccountDetails accountDetails, int exCnt)
+        {
+            if (accountDetails.Owner.ToLower().Equals(accountDetails.Recipient.ToLower()))
+            {
+                ModelState.AddModelError("Recipient", "Recipient email cannot be same as owner email.");
+            }
+            if (CheckAccountRecipientEmail(accountDetails.Recipient, exCnt))
+            {
+                ModelState.AddModelError("Recipient", "Recipient already has an account.");
+            }
+            if (CheckAccountOwnerRecipient(accountDetails.Owner))
+            {
+                ModelState.AddModelError("Owner", "Owner already registered as a recipient.");
+            }
+            if (CheckAccountRecipientOwner(accountDetails.Recipient))
+            {
+                ModelState.AddModelError("Recipient", "Recipient already registered as an owner.");
+            }
         }
     }
 }
